requirements.txt
pandas
pandas-ta
numpy
python-bcb
pyettj
plotly
streamlit
